Crate hog_detector

hog-detector

Histogram of Oriented Gradients and Object Detection

This project contains an algorithm for object detection based on SVM on HOG descriptors.

You can find a demo here

how it works

A training dataset with positive and negative samples is given.

HOG descriptors of the samples are calculated.

A Linear Support Vector Machine is trained on the positive and negative samples.

Using a sliding window, the svm classifier detects bounding boxes.

Applying non-maximum suppression removes some of the bounding boxes.

run examples

You can find examples in example folder. The mnist example loads the mnist dataset and trains hog on the numbers.

cargo run --features mnist --example mnist

run wasm

wasm-pack build --target web
python3 -m http.server

run tests

You can run tests

cargo test

and you can run benchmarks

cargo bench

train data preparation

Find a minimal annotation tool in the demo.

references

https://pyimagesearch.com/2014/11/10/histogram-oriented-gradients-object-detection/

Hard Negative Mining

Eye Dataset

Re-exports

pub use dataset::DataSet;

pub use detector::Detector;

pub use hogdetector::HogDetector;

pub use predictable::Predictable;

pub use trainable::Trainable;

Modules

bbox

bounding box struct and functions

dataset

functionality for loading train datasets

detection

detection struct module

detector

detector functionality

hogdetector

hogdetector struct module

predictable

predict functionality

prelude

the hog_detector prelude

tests

some structs and functions usable for tests

trainable

train functions

utils

some functions for manipulating images

wasm

web assembly module

Type Definitions

Annotation

annotation is a object bounding box in image and class type

Class

object class type