Module hog_detector::feature_descriptor

Feature descriptors are used in computer vision and image processing to describe the features of an image.

Structs

BriefFeatureDescriptor

This code defines a struct to represent a BRIEF (Binary Robust Independent Elementary Features) descriptor. The BriefFeatureDescriptor struct contains a BriefDescriptor struct.

CombinedFeatureDescriptor

defines a feature descriptor combination

HogFeatureDescriptor

This struct is a HogFeatureDescriptor which is used to represent a Histogram of Oriented Features Descriptor. It contains one field, options, which is of type HogOptions.

Traits

FeatureDescriptor

trait for a feature descriptor