Function hog_detector::utils::rotated_frames

pub fn rotated_frames(
    frame: &DynamicImage
) -> impl Iterator<Item = DynamicImage> + '_

rotated_frames() takes in a DynamicImage and returns an iterator over the rotated windows of the given image. It uses rotate_about_center() to rotate the image by a given radian angle, and uses Nearest interpolation with a black background. The radian angles used are 0.02, -0.02, 0.05, -0.05, 0.07, -0.07, 0.09, -0.09, 1.1, -1.1, 1.3, -1.3, 1.5, -1.5 and 2.0,-2.0